requirements.txt
import time
import json
import streamlit as st
from dotenv import load_dotenv
from ideas_engine import generate_hackathon_ideas, refine_top_idea
from schools import schools

load_dotenv()

st.set_page_config(page_title="Hackathon Idea Generator", page_icon="")

st.title("🎯 Your Next Winning Hackathon Idea")
st.write("Generate hackathon project ideas based on past winners in your region.")
if "generated_data" not in st.session_state:
    st.session_state.generated_data = None

if "refined_data" not in st.session_state:
    st.session_state.refined_data = None

if "school_name" not in st.session_state:
    st.session_state.school_name = None

if "location" not in st.session_state:
    st.session_state.location = None

school_name = st.selectbox(
    "Select Your School",
    list(schools.keys())
)

location = schools[school_name]["location"]
st.caption(f"Location: {location}")

interest_area = st.selectbox(
    "Preferred Interest Area",
    ["Open Choice", "AI", "Education", "Healthcare", "Fintech", "Social Good", "Productivity", "Sustainability"]
)

generate_button = st.button("Ranked Project Ideas")

if generate_button:
    with st.spinner("Analyzing past hackathon patterns and generating ideas..."):
        time.sleep(2)
        try:
            result = generate_hackathon_ideas(school_name, location, interest_area)
            data = json.loads(result)
            st.session_state.generated_data = data
            st.session_state.refined_data = None
            st.session_state.school_name = school_name
            st.session_state.location = location 
        except Exception as e:
            st.error(f"Something went wrong: {e}")

if st.session_state.generated_data:
    data = st.session_state.generated_data

    st.subheader("Common Winning Patterns")
    for pattern in data["patterns"]:
        st.write(f"• {pattern}")

    st.subheader("Generated Ideas")

    for i, idea in enumerate(data["ideas"], start=1):
        with st.container():
            if i == 1:
                st.markdown(f"### ⭐ #{i} {idea['name']} (Recommended)")
            else:
                st.markdown(f"### #{i} {idea['name']}")
            st.write(f"**Pitch:** {idea['pitch']}")
            st.write(f"**Why judges may like it:** {idea['why_it_could_rank']}")
            if i == 1:
                st.write(f"🏆**Judge Appeal Score:** {idea['Potential_score']}/10")
            else:
                st.write(f"**Judge Appeal Score:** {idea['Potential_score']}/10")
            difficulty = idea["difficulty"]
            if difficulty == "Easy":
                st.success("🟢 Easy")
            elif difficulty == "Medium":
                st.warning("🟡 Medium")
            else:
                st.error("🔴 Hard")
            st.divider()

    idea_options = list(range(len(data["ideas"])))
    selected_idea_index = st.selectbox(
        "Choose an idea to refine",
        idea_options,
        format_func=lambda i: f"#{i + 1} - {data['ideas'][i]['name']}"
    )

    selected_idea = data["ideas"][selected_idea_index]

    if st.button("Pick an Idea to Refine"):
        with st.spinner(f"Turning {selected_idea['name']} into a realistic hackathon MVP plan..."):
            try:
                refined_result = refine_top_idea(
                    idea_name=selected_idea["name"],
                    idea_pitch=selected_idea["pitch"],
                    school_name=st.session_state.school_name,
                    location=st.session_state.location
                )
                st.session_state.refined_data = json.loads(refined_result)
            except Exception as e:
                st.error(f"Something went wrong while refining the idea: {e}")

    if st.session_state.refined_data:
        refined = st.session_state.refined_data
        st.markdown(f"**Selected Idea:** {selected_idea['name']}")
        st.subheader(f"Refined MVP Plan for {selected_idea['name']}")
        #st.markdown(f"### {refined['mvp_title']}")
        st.write(f"**Core Problem:** {refined['core_problem']}")
        st.write(f"**Target Users:** {refined['target_users']}")

        st.write("**Must-Have Features:**")
        for feature in refined["must_have_features"]:
            st.write(f"• {feature}")

        st.write("**Nice-to-Have Features:**")
        for feature in refined["nice_to_have_features"]:
            st.write(f"• {feature}")

        st.write("**Tech Stack:**")
        for tool in refined["tech_stack"]:
            st.write(f"• {tool}")

        st.write("**How to Actually do it:**")
        for step_num, step in enumerate(refined["implementation_plan"], start=1):
            st.write(f"{step_num}. {step}")

        st.write("**Judge Pitch & How to Frame it:**")
        st.info(refined["judge_pitch"])
